requirements.txt
# requirements.txt en la raíz del repositorio
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.0.52
pydantic>=2.7.0
python-docx>=1.1.0
streamlit>=1.33.0
typing-extensions>=4.11.0
langchain-community>=0.2.0
tavily-python>=0.3.0
python-dotenv>=1.0.0  # Para manejar variables de entorno